Moving KGE functions from KGE_train to sun_models.py
Functions moved:
- save_configs
- save_models
- update_best_model
- clearn_checkpoints

Changes
